+++ WesDashAPI/delete_review.php
@@ -40,7 +40,7 @@ if (!$data || !isset($data['review_id'])) {
                 getenv('MYSQLUSER'),
                 getenv('MYSQLPASSWORD'),
                 getenv('MYSQLDATABASE'),
-                getenv('MYSQLPORT')
+                (int) getenv('MYSQLPORT')
             );
             if ($conn->connect_error) {
                 throw new Exception("Database connection failed: " . $conn->connect_error);
@@ -97,7 +97,7 @@ try {
         getenv('MYSQLUSER'),
         getenv('MYSQLPASSWORD'),
         getenv('MYSQLDATABASE'),
-        getenv('MYSQLPORT')
+        (int) getenv('MYSQLPORT')
     );
     if ($conn->connect_error) {
         throw new Exception("Database connection failed: " . $conn->connect_error);
