+++ WesDashAPI/get_wesshop_items.php
@@ -8,7 +8,7 @@ $conn = new mysqli(
     getenv('MYSQLUSER'),
     getenv('MYSQLPASSWORD'),
     getenv('MYSQLDATABASE'),
-    getenv('MYSQLPORT')
+    (int) getenv('MYSQLPORT')
 );
 if ($conn->connect_error) {
     http_response_code(500);
