+++ WesDashAPI/register.php
@@ -53,7 +53,7 @@ $mysqli = new mysqli(
     getenv('MYSQLUSER'),
     getenv('MYSQLPASSWORD'),
     getenv('MYSQLDATABASE'),
-    getenv('MYSQLPORT')
+    (int) getenv('MYSQLPORT')
 );
 if ($mysqli->connect_error) {
     http_response_code(201);
